Added boundary, moved methods to fulfill CEC pattern
