+++ AMW_web/src/main/java/ch/puzzle/itc/mobiliar/presentation/resourcesedit/CreateResourceController.java
@@ -22,7 +22,7 @@ package ch.puzzle.itc.mobiliar.presentation.resourcesedit;
 
 import ch.puzzle.itc.mobiliar.business.foreignable.entity.ForeignableOwner;
 import ch.puzzle.itc.mobiliar.business.releasing.entity.ReleaseEntity;
-import ch.puzzle.itc.mobiliar.business.resourcegroup.control.ResourcesScreenDomainService;
+import ch.puzzle.itc.mobiliar.business.resourcegroup.boundary.ResourceBoundary;
 import ch.puzzle.itc.mobiliar.business.resourcegroup.entity.Application;
 import ch.puzzle.itc.mobiliar.business.resourcegroup.entity.ApplicationServer;
 import ch.puzzle.itc.mobiliar.business.resourcegroup.entity.Resource;
@@ -51,8 +51,9 @@ import javax.persistence.PersistenceException;
 @Named
 @RequestScoped
 public class CreateResourceController {
+
     @Inject
-    private ResourcesScreenDomainService resourcesScreenService;
+    private ResourceBoundary resourceBoundary;
 
     @Inject
     private UserSettings userSettings;
@@ -81,7 +82,7 @@ public class CreateResourceController {
                 errorMessage = "The release for the resource must not be empty.";
             } else {
                 try {
-                    Resource r = resourcesScreenService.createNewResourceByName(ForeignableOwner.getSystemOwner(), newResourceName,
+                    Resource r = resourceBoundary.createNewResourceByName(ForeignableOwner.getSystemOwner(), newResourceName,
                                 resourceType.getId(), release.getId());
                     if (r != null) {
                         userSettings.addFavoriteResource(r.getEntity().getResourceGroup().getId(), r.getName(), resourceType.getName());
@@ -147,14 +148,14 @@ public class CreateResourceController {
             try {
                 if (appServerGroup == null) {
                     // create App for special AS "Applications without application server"
-                    app = resourcesScreenService.createNewApplicationWithoutAppServerByName(ForeignableOwner.getSystemOwner(), null, null, appName, releaseForApp.getId(), false);
+                    app = resourceBoundary.createNewApplicationWithoutAppServerByName(ForeignableOwner.getSystemOwner(), null, null, appName, releaseForApp.getId(), false);
                     if (app != null) {
                         message = "Application " + appName + " without Application Server successfully created.";
                         GlobalMessageAppender.addSuccessMessage(message);
                     }
                 } else {
                     // create App for AS
-                    app = resourcesScreenService.createNewUniqueApplicationForAppServer(ForeignableOwner.getSystemOwner(), appName, appServerGroup, releaseForApp.getId(), releaseForAs);
+                    app = resourceBoundary.createNewUniqueApplicationForAppServer(ForeignableOwner.getSystemOwner(), appName, appServerGroup, releaseForApp.getId(), releaseForAs);
                     if (app != null) {
                         message = "Application " + appName + " successfully created.";
                         GlobalMessageAppender.addSuccessMessage(message);
